Add support for 386(32bit) and ARM(32bit).
This is required to be able to compile minio server for multiple
architecture builds.
